# frontend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Modular-monolith manufacturing ERP — double-entry ledger, inventory, procure-to-pay, order-to-cash, manufacturing and reporting.",
    "title" : "Manufacturing ERP API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/auth/login" : {
      "post" : {
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/auth/refresh" : {
      "post" : {
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "cookie",
          "name" : "refresh_token",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "auth-controller" ]
      }
    },
...
    "/api/hr/leave-requests/{id}/approve" : {
      "post" : {
        "operationId" : "approve_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LeaveRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "leave-controller" ]
      }
    },
...
    "/api/hr/timesheets" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "employeeId",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "DRAFT", "SUBMITTED", "APPROVED" ]
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TimesheetResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "timesheet-controller" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTimesheetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimesheetResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "timesheet-controller" ]
      }
    },
    "/api/hr/timesheets/{id}/approve" : {
      "post" : {
        "operationId" : "approve",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimesheetResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "timesheet-controller" ]
      }
    },
    "/api/hr/timesheets/{id}/submit" : {
      "post" : {
        "operationId" : "submit",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimesheetResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "timesheet-controller" ]
      }
    },
    "/api/inventory/adjustments" : {
      "post" : {
        "operationId" : "postAdjustment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAdjustmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
...
    "/api/inventory/items/{itemId}/on-hand" : {
      "get" : {
        "operationId" : "onHand",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ItemStock"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
...
    "/api/reporting/reconciliation" : {
      "get" : {
        "operationId" : "reconciliation",
        "parameters" : [ {
          "in" : "query",
          "name" : "asOf",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "reporting-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AdjustmentResponse" : {
        "type" : "object",
        "properties" : {
          "adjustmentNumber" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "journalEntryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "locationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "movementGroupId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "postedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "qtyDelta" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "unitCost" : {
            "type" : "string"
          }
        }
      },
...
      "ClearingBalance" : {
        "type" : "object",
        "properties" : {
          "accountCode" : {
            "type" : "string"
          },
          "balance" : {
            "type" : "string"
          },
          "cleared" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "CreateAdjustmentRequest" : {
        "type" : "object",
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "locationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "qtyDelta" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "unitCost" : {
            "type" : "string"
          }
        },
        "required" : [ "itemId", "locationId", "qtyDelta" ]
      },
...
      "CreateTimesheetRequest" : {
        "type" : "object",
        "properties" : {
          "employeeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "note" : {
            "type" : "string"
          },
          "overtimeHours" : {
            "type" : "string"
          },
          "regularHours" : {
            "type" : "string"
          },
          "weekEnding" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "employeeId", "overtimeHours", "regularHours", "weekEnding" ]
      },
...
      "ItemStock" : {
        "type" : "object",
        "properties" : {
          "avgUnitCost" : {
            "type" : "string"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "onHandQty" : {
            "type" : "string"
          },
          "sku" : {
            "type" : "string"
          },
          "totalValue" : {
            "type" : "string"
          }
        }
      },
...
      "LeaveRequestResponse" : {
        "type" : "object",
        "properties" : {
          "days" : {
            "type" : "string"
          },
          "decidedBy" : {
            "type" : "string"
          },
          "decidedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "employeeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "leaveType" : {
            "type" : "string",
            "enum" : [ "ANNUAL", "SICK", "PERSONAL", "UNPAID" ]
          },
          "reason" : {
            "type" : "string"
          },
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "APPROVED", "REJECTED" ]
          }
        }
      },
...
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "password" : {
            "type" : "string",
            "minLength" : 1
          },
          "username" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "password", "username" ]
      },
...
      "ReconciliationReport" : {
        "type" : "object",
        "properties" : {
          "asOf" : {
            "type" : "string",
            "format" : "date"
          },
          "clearingAccounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ClearingBalance"
            }
          },
          "healthy" : {
            "type" : "boolean"
          },
          "subledgerChecks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubledgerCheck"
            }
          },
          "trialBalanceBalanced" : {
            "type" : "boolean"
          }
        }
      },
...
      "SubledgerCheck" : {
        "type" : "object",
        "properties" : {
          "accountCode" : {
            "type" : "string"
          },
          "glControl" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "reconciled" : {
            "type" : "boolean"
          },
          "subledger" : {
            "type" : "string"
          }
        }
      },
...
      "TimesheetResponse" : {
        "type" : "object",
        "properties" : {
          "employeeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "note" : {
            "type" : "string"
          },
          "overtimeHours" : {
            "type" : "string"
          },
          "regularHours" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "SUBMITTED", "APPROVED" ]
          },
          "weekEnding" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "TokenResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "username" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}